+++ ios/Flutter/flutter_export_environment.sh
@@ -3,6 +3,7 @@
 export "FLUTTER_ROOT=/opt/homebrew/share/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/gabrielriccimazuqueli/Documents/Projetos/Ricci/Flutter/AwesomeAlert"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
+export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.1.6"
 export "FLUTTER_BUILD_NUMBER=1.1.6"